--- doc/en/user/source/community/authbasic/install.rst
@@ -1,11 +1,0 @@
-.. _authbasic_install:
-
-Installing the HTTP Based Authorization plug-in
-===============================================
-
-#. Download the ``gs-authbasic`` plug-in from the `nightly GeoServer community module builds <https://build.geoserver.org/geoserver/main/community-latest/>`_.
-
-    .. warning:: Make sure to match the version of the extension to the version of the GeoServer instance!
-
- #. Extract the contents of the archive into the ``WEB-INF/lib`` directory of the GeoServer installation.
-
