+++ doc/en/user/source/community/index.rst
@@ -60,4 +60,4 @@ officially part of the GeoServer releases. They are however built along with the
    cov-json/index
    smart-data-loader/index
    schemaless-features/index
-   authbasic/index
+   web-service-auth/index
