+++ doc/en/user/source/community/web-service-auth/install.rst
@@ -1,0 +1,11 @@
+.. _webserviceauth_install:
+
+Installing the HTTP Based Authorization plug-in
+===============================================
+
+#. Download the ``gs-web-service-auth`` plug-in from the `nightly GeoServer community module builds <https://build.geoserver.org/geoserver/main/community-latest/>`_.
+
+    .. warning:: Make sure to match the version of the extension to the version of the GeoServer instance!
+
+ #. Extract the contents of the archive into the ``WEB-INF/lib`` directory of the GeoServer installation.
+
