--- src/community/authbasic/src/main/java/applicationContext.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE beans PUBLIC "-//SPRING//DTD BEAN//EN" "http://www.springframework.org/dtd/spring-beans.dtd">
-
-<beans>
-   <bean class="org.geoserver.platform.ModuleStatusImpl">
-    <constructor-arg index="0" value="authbasic"/>
-    <constructor-arg index="1" value="HTTP Based Authorization"/>
-  </bean>
-   
-   <bean id="simpleWebAuthProviderPanelInfo" class="org.geoserver.security.web.auth.SimpleWebAuthProviderPanelInfo">
-        <property name="id" value="security.simpleWebAuthProvider" />
-        <property name="shortTitleKey" value="SimpleWebAuthProviderPanel.short"/>
-        <property name="titleKey" value="SimpleWebAuthProviderPanel.title"/>
-        <property name="descriptionKey" value="SimpleWebAuthProviderPanel.description"/>
-    </bean>
-</beans>
