--- src/community/authbasic/src/main/resources/GeoServerApplication.properties
@@ -1,15 +1,0 @@
-org.geoserver.security.auth.web.SimpleWebServiceAuthenticationProvider.name=Web Service Authentication
-org.geoserver.security.auth.web.SimpleWebServiceAuthenticationProvider.title=Authenticate using Web Service
-
-SimpleWebAuthProviderPanel.short=Web Service Authentication
-SimpleWebAuthProviderPanel.title=Web Service Authentication
-SimpleWebAuthProviderPanel.description=Use an external web service to authenticate credentials
-SimpleWebAuthProviderPanel.connectionURL= Service URL
-SimpleWebAuthProviderPanel.readTimeoutOut= Read Timeout
-SimpleWebAuthProviderPanel.connectionTimeOut= Timeout
-SimpleWebAuthProviderPanel.roleRegex= Regex Expression for Role Parsing (Optional)
-SimpleWebAuthProviderPanel.roleService= Role Service
-SimpleWebAuthProviderPanel.useHeader= Send credentials in X-HTTP-AUTHORIZATION Header (user:password)
-SimpleWebAuthProviderPanel.connectionSettingsLegend= Web Service Setting
-SimpleWebAuthProviderPanel.authorizationSettingsLegend= Authorization
-SimpleWebAuthProviderPanel.allowHTTPConnection= Allow HTTP connection
