--- src/community/release/ext-authbasic.xml
@@ -1,16 +1,0 @@
-<assembly>
-  <id>authbasic</id>
-  <formats>
-    <format>zip</format>
-  </formats>
-  <includeBaseDirectory>false</includeBaseDirectory>
-  <fileSets>
-    <fileSet>
-      <directory>release/target/dependency</directory>
-      <outputDirectory></outputDirectory>
-      <includes>
-        <include>gs-authbasic*.jar</include>
-      </includes>
-    </fileSet>
-  </fileSets>
-</assembly>
