+++ src/community/release/pom.xml
@@ -408,7 +408,7 @@
     </dependency>
     <dependency>
       <groupId>org.geoserver.community</groupId>
-      <artifactId>gs-authbasic</artifactId>
+      <artifactId>gs-web-service-auth</artifactId>
       <version>${project.version}</version>
     </dependency>
   </dependencies>
