+++ src/community/web-service-auth/src/main/java/applicationContext.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE beans PUBLIC "-//SPRING//DTD BEAN//EN" "http://www.springframework.org/dtd/spring-beans.dtd">
+
+<beans>
+   <bean class="org.geoserver.platform.ModuleStatusImpl">
+    <constructor-arg index="0" value="authbasic"/>
+    <constructor-arg index="1" value="HTTP Based Authorization"/>
+  </bean>
+   
+   <bean id="webAuthProviderPanelInfo" class="org.geoserver.security.web.auth.WebAuthProviderPanelInfo">
+        <property name="id" value="security.webAuthProvider" />
+        <property name="shortTitleKey" value="WebAuthProviderPanel.short"/>
+        <property name="titleKey" value="WebAuthProviderPanel.title"/>
+        <property name="descriptionKey" value="WebAuthProviderPanel.description"/>
+    </bean>
+</beans>
