+++ src/community/web-service-auth/src/main/resources/GeoServerApplication.properties
@@ -1,0 +1,15 @@
+org.geoserver.security.auth.web.WebServiceAuthenticationProvider.name=Web Service Authentication
+org.geoserver.security.auth.web.WebServiceAuthenticationProvider.title=Authenticate using Web Service
+
+WebAuthProviderPanel.short=Web Service Authentication
+WebAuthProviderPanel.title=Web Service Authentication
+WebAuthProviderPanel.description=Use an external web service to authenticate credentials
+WebAuthProviderPanel.connectionURL= Service URL
+WebAuthProviderPanel.readTimeoutOut= Read Timeout
+WebAuthProviderPanel.connectionTimeOut= Timeout
+WebAuthProviderPanel.roleRegex= Regex Expression for Role Parsing (Optional)
+WebAuthProviderPanel.roleService= Role Service
+WebAuthProviderPanel.useHeader= Send credentials in X-HTTP-AUTHORIZATION Header (user:password)
+WebAuthProviderPanel.connectionSettingsLegend= Web Service Setting
+WebAuthProviderPanel.authorizationSettingsLegend= Authorization
+WebAuthProviderPanel.allowHTTPConnection= Allow HTTP connection
