+++ src/community/web-service-auth/src/main/resources/GeoServerException.properties
