+++ src/web/app/pom.xml
@@ -1796,11 +1796,11 @@
       </dependencies>
     </profile>
    <profile>
-      <id>authbasic</id>
+      <id>web-service-auth</id>
       <dependencies>
         <dependency>
           <groupId>org.geoserver.community</groupId>
-          <artifactId>gs-authbasic</artifactId>
+          <artifactId>gs-web-service-auth</artifactId>
           <version>${gs.version}</version>
         </dependency>
       </dependencies>
